refactor(skills): compose skill types and share the session catalog contract
Follow-up to the review discussion on the SkillCatalogEntry name collision.
Both changes are type-only; the emitted JavaScript is unchanged.

SkillCandidate, Skill, and SkillCatalogEntry each repeated the same 23 field
lines verbatim, so adding a skill field meant editing three declarations and
missing one failed silently. Skill and SkillCatalogEntry now extend
SkillCandidate with only their own delta, matching what
BlockedSkillCatalogEntry already did, and each carries a doc comment naming
the surface it serves.

SessionSkillCatalogEntry was declared twice, once per side of the host and
container boundary, leaving the wire contract synced by hand. It now lives in
container/shared/skill-catalog.d.ts and is imported and re-exported by both
type modules, the same way WebSearchConfig is already shared.
